requirements.txt
streamlit
requests
pillow
urllib3
